refactor(config): extract TomlEditor from ConfigManager
Move all surgical toml_edit operations into a new TomlEditor struct
in src/config/toml_editor.rs.  ConfigManager now holds a TomlEditor
field and delegates every sandbox-key method to it, keeping the
public API identical.  No behavioral change — pure structural extraction.
